Changes in user/serializers.py and smmart/serializers.py
